pyproject.toml
[project]
name = "programador-citas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "sib-api-v3-sdk>=7.6.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]
